+++ src/lib/components/Currency/Currency.js
@@ -1,6 +1,9 @@
 import PropTypes from 'prop-types';
 
 export default function Currency({ value }) {
+  if (isNaN(value)) {
+    return null;
+  }
   const options = {
     minimumFractionDigits: 2,
     style: 'currency',