+++ webpack.config.publish.js
@@ -8,7 +8,13 @@ module.exports = {
   entry: path.resolve(SRC_DIR, 'index.js'),
   output: {
     path: DIST_DIR,
-    filename: 'bundle.js'
+    filename: 'bundle.js',
+    library: {
+      root: "IncreaseComponents",
+      amd: "increase-components",
+      commonjs: "increase-components"
+    },
+    libraryTarget: 'umd'
   },
   module: {
     loaders: [